Remove resources classes from Feature implementation
Registering endpoints in a Feature is a gray area with JAX-RS. This seems to work with Jetty but not Wildfly.
Based on community feedback this doesn't seem to be a valid expectation of a Feature.

This change limits ScimpleFeature to only registering ExceptionMappers and Media Type related classes
And adds util functionality in ScimResourceHelper