Fixing return type of fluent setters extending BaseResource
NOTE: It's possible to use generics in the definition of the ScimResource, but this complicates the use of ScimResource in other places (specifically the bulk resource)
I've hacked on this a couple different times and each time I've reverted a bunch of ugly code.
At the end of the day the generic type usage is only to facilitate the fluent setters. It seems less complicated to override the setters for this purpose.
Ideally the use of lombok would handle this automatically.