Move @Context object to field injection
Previously all injectsion into resources were done with field injection, this was replaced with constructor injection and the RequestContext class.
However this mixes CDI injection and the Jakarta WS injection techniques, not all implementations support this.

Jakarta WS 4 should allow these objects to be injected via CDI-light.

For now there is a mix of constructor and field injection with a TODO message where they are used.