Pass all search information into the repository implementation
This change is backward compatible, but it's possible this should be a breaking change, and force implementations to use this method, instead of the other?
When developer overrides this method, the original `find` method must still be defined, but is not used.

Questions:
-  I'm not really a fan of the FindRequest nameing, `SearchRequest` or `QueryRequest` might work better, but `SearchRequest` is used at the REST layer, so I didn't thing we should have two classes named `SearchRequest` (maybe we should rename the other...)
- Backwards compatibilty, should we break this? (the migration guide would be simple)
- Passing the exclude args down the the search is probably the right thing to do, but that potentially changes the meaning of the etag,  this is currently done at the REST layer, but maybe this should move down a layer?  We could still handle this by default 🤷

We probably need to dig into how the etag should work when attributes are included/excluded