Replaced zjsonpatch based resource diffing with Attribute based solution
The json patch solution required seralizing objects, then processing them with zjsonpatch, and finally some heavy post processing.

This new solution walks the tree schema attributes to calculate the diff directly.
Class was moved to scim-tools, as it's not directly used by scim-server or scim-client.
PatchGenerator could be used by a SCIM client to calculate diffs to send to a server
Or to generate patch requests when testing a server.

NOTE: There are some limitations, this implementation assumes any changes to an object in a collection modifies the whole object,
this will alwasy trigger a REMOVE and ADD operation (instead of a REPLACE).